changing file structure by creating abstract_equipment_classes
